+++ cassandra/src/main/scala/akka/stream/alpakka/cassandra/CassandraSourceStage.scala
@@ -6,12 +6,11 @@ package akka.stream.alpakka.cassandra
 import akka.stream._
 import akka.stream.stage.{ AsyncCallback, GraphStage, GraphStageLogic, OutHandler }
 import com.datastax.driver.core.{ ResultSet, Row, Session, Statement }
-import com.google.common.util.concurrent.{ FutureCallback, Futures, ListenableFuture }
 
-import scala.concurrent.{ Future, Promise }
+import scala.concurrent.Future
 import scala.util.{ Failure, Success, Try }
 
-import GuavaFutureOptsImplicits._
+import akka.stream.alpakka.cassandra.cassandra._
 
 class CassandraSourceStage(futStmt: Future[Statement], session: Session) extends GraphStage[SourceShape[Row]] {
   val out: Outlet[Row] = Outlet("CassandraSource.out")
@@ -27,7 +26,7 @@ class CassandraSourceStage(futStmt: Future[Statement], session: Session) extends
 
         futFetchedCallback = getAsyncCallback[Try[ResultSet]](tryPushAfterFetch)
 
-        val futRs = futStmt.flatMap(stmt => session.executeAsync(stmt).toFuture())
+        val futRs = futStmt.flatMap(stmt => session.executeAsync(stmt).asScala())
         futRs.onComplete(futFetchedCallback.invoke)
       }
 
@@ -40,7 +39,7 @@ class CassandraSourceStage(futStmt: Future[Statement], session: Session) extends
             case Some(rs) if rs.isExhausted                     => completeStage()
             case Some(rs) =>
               // fetch next page
-              val futRs = rs.fetchMoreResults().toFuture()
+              val futRs = rs.fetchMoreResults().asScala()
               futRs.onComplete(futFetchedCallback.invoke)
             case None => () // doing nothing, waiting for futRs in preStart() to be completed
           }
